+++ include/groot/cgal.hpp
@@ -13,7 +13,7 @@
 
 namespace groot::cgal {
 
-typedef CGAL::Simple_cartesian<float> Kernel;
+typedef CGAL::Simple_cartesian<double> Kernel;
 
 typedef Kernel::Point_3 Point_3;
 typedef Kernel::Vector_3 Vector_3;
